Make OpenJDK 17 build hermetic
The OpenJDK 11 build that the OpenJDK 17 build was copied from
was not hermetic, it was reusing glibc from the host.  That resulted
in a java binary that would not run on hosts with an older version
of glibc.  Unpack all the debs and add the --sysroot flag to cflags,
cxxflags and ldflags so that glibc always comes from the debs, which
will ensure compatibility with glibc 2.19.

Bug: 219098645
Test: toolchain/jdk/build/build-openjdk17-linux.sh
Change-Id: Ie59303f097840a308db6d3e47098b734ff88732f