add details bundled with JBR release file
Details added include SOURCE, IMPLEMENTOR, OS_NAME, and OS_ARCH.

This is to address a bug caused by IDE looking for the OS_ARCH
in the release file before selecting the JetBrainsRuntime
packaged with the IDE on startup for Linux.

Additionally, begin distributing the 'release' file to make
seeing which version is distributed an easier process.
Fixes: 244373539

Change-Id: I9468e30c0394bf0c70c3f63f9ea4cd30c6827929